Added README, removed style-guide
A README file was added, so you know what and why this project exists and the style-guide was removed, because it just contained some information on what colors and fonts to use for this challenge. You can find them on the frontendmentor.io website if you would like to participate.